% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/require.R
\name{library}
\alias{library}
\alias{require}
\title{Loading/Attaching Packages}
\usage{
library(package, help, pos = 2, lib.loc = NULL, character.only = FALSE,
  logical.return = FALSE, warn.conflicts = TRUE, quietly = FALSE,
  verbose = getOption("verbose"))

require(package, lib.loc = NULL, quietly = FALSE, warn.conflicts = TRUE,
  character.only = FALSE)
}
\arguments{
\item{package,help}{the name of a package, given as a name or literal character string,
or a character string, depending on whether character.only is FALSE (default) or TRUE).}

\item{pos}{the position on the search list at which to attach the loaded namespace.
Can also be the name of a position on the current search list as given by search().}

\item{lib.loc}{a character vector describing the location of R library trees to search
through, or NULL. The default value of NULL corresponds to all libraries currently known
to .libPaths(). Non-existent library trees are silently ignored.}

\item{character.only}{a logical indicating whether package or help can be assumed to
be character strings.}

\item{logical.return}{logical. If it is TRUE, FALSE or TRUE is returned to indicate success.}

\item{warn.conflicts}{logical. If TRUE, warnings are printed about conflicts from attaching
the new package. A conflict is a function masking a function, or a non-function masking
a non-function or a method masking a method.}

\item{quietly}{a logical. If TRUE, no message confirming package attaching is printed,
and most often, no errors/warnings are printed if package attaching fails.}

\item{verbose}{a logical. If TRUE, additional diagnostics are printed.}
}
\description{
Same functions as \code{library} and \code{require} from base with
improved warnings when overloading S3 methods. These functions use
the underlying base functions together with \code{\link{warnS3Methods}}
}
\examples{
\dontrun{
require(dplyr)
library(roxygen2)
}
}
\seealso{
\code{\link[base]{library}}, \code{\link[base]{require}}
}

