% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ns-export.R
\name{ns-dblcolon}
\alias{::}
\alias{:::}
\alias{ns-dblcolon}
\title{Double Colon and Triple Colon Operators}
\usage{
"::"(pkg, name)

":::"(pkg, name)
}
\arguments{
\item{pkg}{package name}

\item{name}{variable name}
}
\description{
Accessing exported and internal variables in a namespace
with a warning when loading the namespace.
}

