% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/warnings.R
\name{warnS3Methods}
\alias{warnS3Methods}
\title{Warn if package overloads S3 methods}
\usage{
warnS3Methods(pkg = NULL)
}
\arguments{
\item{pkg}{character of package to ckeck. If NULL, check all loaded packages.}
}
\description{
Warn if a package overloads S3 methods from other loaded packages
}
\examples{
warnS3Methods()
}

